pyproject.toml
[tool.poetry]
name = "fastapi-docker"
version = "0.1.0"
description = ""
authors = ["saqlain"]
license = "n"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
